Add files via upload
To monitor the impact of COVID-19 vaccinations, a death count and population count time and location tracker can be created, with SQL queries used to retrieve data on the number of deaths and vaccinations administered globally. This information can be used to identify trends and hotspots, and to target vaccination efforts and resources to areas with the highest need.